+++ src/room/track/Track.ts
@@ -121,7 +121,9 @@ export abstract class Track extends (EventEmitter as new () => TypedEventEmitter
     // we'll want to re-attach it in that case
     attachToElement(this._mediaStreamTrack, element);
 
-    if (element instanceof HTMLAudioElement) {
+    // handle auto playback failures
+    const allMediaStreamTracks = (element.srcObject as MediaStream).getTracks();
+    if (allMediaStreamTracks.some((tr) => tr.kind === 'audio')) {
       // manually play audio to detect audio playback status
       element
         .play()
@@ -130,6 +132,17 @@ export abstract class Track extends (EventEmitter as new () => TypedEventEmitter
         })
         .catch((e) => {
           this.emit(TrackEvent.AudioPlaybackFailed, e);
+          // If audio playback isn't allowed make sure we still play back the video
+          if (
+            element &&
+            allMediaStreamTracks.some((tr) => tr.kind === 'video') &&
+            e.name === 'NotAllowedError'
+          ) {
+            element.muted = true;
+            element.play().catch(() => {
+              // catch for Safari, exceeded options at this point to automatically play the media element
+            });
+          }
         });
     }
 
@@ -259,6 +272,13 @@ export function attachToElement(track: MediaStreamTrack, element: HTMLMediaEleme
     mediaStream.addTrack(track);
   }
 
+  element.autoplay = true;
+  // In case there are no audio tracks present on the mediastream, we set the element as muted to ensure autoplay works
+  element.muted = mediaStream.getAudioTracks().length === 0;
+  if (element instanceof HTMLVideoElement) {
+    element.playsInline = true;
+  }
+
   // avoid flicker
   if (element.srcObject !== mediaStream) {
     element.srcObject = mediaStream;
@@ -279,10 +299,6 @@ export function attachToElement(track: MediaStreamTrack, element: HTMLMediaEleme
         });
       }, 0);
     }
-  }
-  element.autoplay = true;
-  if (element instanceof HTMLVideoElement) {
-    element.playsInline = true;
   }
 }